fetches network name in nav-host
