+++ app/app/src/main/java/com/bringyour/network/ui/MainNavHost.kt
@@ -7,24 +7,48 @@ import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.padding
 import androidx.compose.material3.Scaffold
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.LaunchedEffect
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.tooling.preview.Preview
-import androidx.hilt.navigation.compose.hiltViewModel
+import androidx.compose.ui.platform.LocalContext
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import androidx.navigation.compose.rememberNavController
+import com.bringyour.network.MainApplication
 import com.bringyour.network.ui.components.BottomNavBar
 import com.bringyour.network.ui.components.overlays.FullScreenOverlay
 import com.bringyour.network.ui.connect.ConnectScreen
 import com.bringyour.network.ui.connect.ConnectViewModel
 import com.bringyour.network.ui.feedback.FeedbackScreen
-import com.bringyour.network.ui.theme.URNetworkTheme
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.runBlocking
 
 
 @Composable
 fun MainNavHost(
     connectViewModel: ConnectViewModel
 ) {
+
+    val context = LocalContext.current
+    val application = context.applicationContext as? MainApplication
+    var networkName by remember { mutableStateOf<String?>(null) }
+
+    val populateNetworkName = {
+        application?.asyncLocalState?.parseByJwt { byJwt, ok ->
+            runBlocking(Dispatchers.Main.immediate) {
+                if (ok) {
+                    networkName = byJwt.networkName
+                }
+            }
+        }
+    }
+
+    LaunchedEffect(Unit) {
+        populateNetworkName()
+    }
 
     val navController = rememberNavController()
 
@@ -51,7 +75,10 @@ fun MainNavHost(
                 exitTransition = { ExitTransition.None }
             ) {
                 composable("connect") {
-                    ConnectScreen(connectViewModel)
+                    ConnectScreen(
+                        connectViewModel,
+                        networkName
+                    )
                 }
                 composable("account") { AccountNavHost() }
                 composable("support") { FeedbackScreen() }
@@ -61,11 +88,3 @@ fun MainNavHost(
         FullScreenOverlay()
     }
 }
-
-//@Preview
-//@Composable
-//private fun MainNavHostPreview() {
-//    URNetworkTheme {
-//        MainNavHost()
-//    }
-//}
